+++ ChangeLog.md
@@ -3,6 +3,7 @@
     - Improved the avatar upload modal (now way more intuitive)
     - Fixed a bug which cause client avatars to be stuck within the loading screen
     - Don't spam permission errors if we don't have the permission to view the channel description
+    - Showing channel group inheritance within the client info frame
   
 * **23.03.21**
     - Made the permission editor popoutable